[BE#35] refactors filtering
