+++ server/services/counselling-service.service.ts
@@ -1,4 +1,5 @@
 import { CounsellingService, ICounsellingService } from '../models/counsellingService.model';
+import { getFilter } from "../utils/filter.util";
 
 async function getCounsellingServices(serviceNameQuery: any,
                                       locationQuery: any,
@@ -13,17 +14,17 @@ async function getCounsellingServices(serviceNameQuery: any,
                                       descriptionQuery: any,
                                       searchString: any): Promise<ICounsellingService[]> {
 
-  const serviceName = await getFilter("serviceName", serviceNameQuery);
-  const location = await getFilter("location", locationQuery);
-  const school = await getFilter("school", schoolQuery);
-  const organization = await getFilter("organization", organizationQuery);
-  const serviceType = await getFilter("serviceType", serviceTypeQuery);
-  const specialty = await getFilter("specialty", specialtyQuery);
-  const urgency = await getFilter("urgency", urgencyQuery);
-  const targetClients = await getFilter("targetClients", targetClientsQuery);
+  const serviceName = getFilter("serviceName", serviceNameQuery);
+  const location = getFilter("location", locationQuery);
+  const school = getFilter("school", schoolQuery);
+  const organization = getFilter("organization", organizationQuery);
+  const serviceType = getFilter("serviceType", serviceTypeQuery);
+  const specialty = getFilter("specialty", specialtyQuery);
+  const urgency = getFilter("urgency", urgencyQuery);
+  const targetClients = getFilter("targetClients", targetClientsQuery);
   const isAllDay = (isAllDayQuery != null) ? { isAllDay: isAllDayQuery } : {};
-  const delivery = await getFilter("delivery", deliveryQuery);
-  const description = await getFilter("description", descriptionQuery);
+  const delivery = getFilter("delivery", deliveryQuery);
+  const description = getFilter("description", descriptionQuery);
 
   const search = searchString ? { $text: { $search: searchString } } : {};
 
@@ -35,25 +36,6 @@ async function getCounsellingServices(serviceNameQuery: any,
                                            .lean();
 
   return services;
-}
-
-async function getFilter(key: string, query: any): Promise<any> {
-    const regExp : RegExp[] = [];
-    if (query && Array.isArray(query)) {
-        (query as string[]).forEach(function(opt: string) {
-            regExp.push( new RegExp(opt, "i") );
-        });
-    }
-
-    const filter = query ?
-        regExp.length == 0 ?
-            { [key]: {$regex: query, $options: 'i'} }
-            :
-            { [key]: {$in: regExp} }
-        :
-        {};
-
-    return await filter;
 }
 
 async function getCounsellingService(id: string): Promise<ICounsellingService> {
